feat(mvvm): Actually use MVVM practices
The whole point of MVVM is to not explicitly refer to a view from the view model layer. The view and view model communicate via properties and commands.

What this means is that you have a separation of concerns between your views and your application logic.

This generically means, "Never Use Find Control".

Doing this also means you can kill most of your scheduling code, because you're not actually touching the UI thread, Avalonia's binding system is doing the marhsaling between UI and non-ui code for you.